Minor changes to visual-semantics in layout and behavior of view2 following certain conventions.
Changed the value property of counter-element in my-view2.js to reflect props._value.

It's the value that the end-user is likely to be interested in, and clicks are not usually relevant as a visual quantity.

The (+) and (-) buttons should produce a distinctly visible change to be more effective i.e. the (-) button should decrement something, and since the number of clicks cannot be decremented, it counter-intuitively seems to do nothing.

Removed repeated display of the value near the buttons, as it's unnecessarily confusing. It makes sense that the input area shows input-related data and the main display area shows the stored data, thereby also reinforcing the feeling that the data is being updated somewhere else, namely the redux store.

Changed the body text to reflect the changes.

Swapped the order of the (+) and (-) buttons, as conventionally the right-side is visually associated with 'positive' and 'increment'.

Changed the order of the dispatches to 'increment' and 'decrement' in my-view2.js to reflect the button-order change.

Caveats: view3 needs to be re-thought/organised if similar conventions are followed.